requirements.txt
Django>=4.2,<5
psycopg[binary]>=3.1,<4
python-dotenv>=1.0
djangorestframework>=3.15,<4
reportlab>=4.2,<5
